change __init__.py
